Extract IssueBoard into reusable library
