+++ libs/vs-issue-board/src/vs_issue_board/__init__.py
@@ -1,0 +1,28 @@
+from vs_issue_board.core import (
+    Issue,
+    IssueBoard,
+    IssueEvent,
+    IssueStatus,
+    IssueType,
+)
+from vs_issue_board.format import format_issue_full, format_issue_short
+from vs_issue_board.policy import (
+    CreateIssuePolicy,
+    check_create_allowed,
+    create_issue_under_policy,
+    parse_type,
+)
+
+__all__ = [
+    "CreateIssuePolicy",
+    "Issue",
+    "IssueBoard",
+    "IssueEvent",
+    "IssueStatus",
+    "IssueType",
+    "check_create_allowed",
+    "create_issue_under_policy",
+    "format_issue_full",
+    "format_issue_short",
+    "parse_type",
+]
