+++ libs/vs-issue-board/src/vs_issue_board/py.typed
@@ -1,0 +1,1 @@
+
