+++ pyproject.toml
@@ -25,11 +25,12 @@ dependencies = [
     "librosa>=0.11.0",
     "modal>=1.4.2",
     "vs-feature-flags",
+    "vs-issue-board",
 ]
 
 [project.scripts]
 vibe-serve = "vibe_serve.cli:main"
-vibe-serve-issue-mcp = "vibe_serve.loops.plain.mcp_server:main"
+vibe-serve-issue-mcp = "vs_issue_board.mcp:main"
 
 [tool.setuptools.packages.find]
 where = ["src"]
@@ -53,14 +54,14 @@ test = [
 ]
 
 [tool.pytest.ini_options]
-testpaths = ["tests", "libs/vs-feature-flags/tests"]
+testpaths = ["tests", "libs/vs-feature-flags/tests", "libs/vs-issue-board/tests"]
 addopts = "-ra"
 pythonpath = ["."]
 
 [tool.ruff]
 line-length = 100
 target-version = "py311"
-src = ["src", "tests", "libs/vs-feature-flags/src", "libs/vs-feature-flags/tests"]
+src = ["src", "tests", "libs/vs-feature-flags/src", "libs/vs-feature-flags/tests", "libs/vs-issue-board/src", "libs/vs-issue-board/tests"]
 
 [tool.ruff.lint]
 select = ["E", "F", "I", "UP", "B", "W"]
@@ -76,3 +77,4 @@ members = ["libs/*"]
 [tool.uv.sources]
 agentshim = { git = "https://github.com/vic-lsh/agentshim.git", rev = "codex/executor-injection" }
 vs-feature-flags = { workspace = true }
+vs-issue-board = { workspace = true }
