+++ tests/loops/agent/test_issue_tools.py
@@ -1,7 +1,7 @@
 """Tests for the issue-tracker tools (list/get/search/create)."""
 
-from vibe_serve.loops.plain.issue_board import IssueBoard, IssueStatus, IssueType
 from vibe_serve.loops.plain.tools import build_issue_tools
+from vs_issue_board import IssueBoard, IssueStatus, IssueType
 
 
 def _make_store(tmp_path) -> IssueBoard:
