+++ tests/test_agent_cli_mcp.py
@@ -33,7 +33,7 @@ def _spec() -> MCPServerSpec:
         command="python",
         args=[
             "-m",
-            "vibe_serve.loops.plain.mcp_server",
+            "vs_issue_board.mcp",
             "issues.json",
             "--creator",
             "judge",
@@ -51,7 +51,7 @@ def _spec_with_env() -> MCPServerSpec:
     return MCPServerSpec(
         name="vibeserve-issues",
         command="python",
-        args=["-m", "vibe_serve.loops.plain.mcp_server", "issues.json"],
+        args=["-m", "vs_issue_board.mcp", "issues.json"],
         env={"MY_VAR": "my_value", "OTHER": "x"},
     )
 
@@ -245,7 +245,7 @@ class TestCodexMCP:
         args_value = args_entries[0].split("=", 1)[1]
         assert args_value.startswith("[") and args_value.endswith("]")
         assert '"-m"' in args_value
-        assert '"vibe_serve.loops.plain.mcp_server"' in args_value
+        assert '"vs_issue_board.mcp"' in args_value
         assert '"--creator"' in args_value
         assert '"judge"' in args_value
         assert '"--cap"' in args_value
